Forbes Richest Atheletes EDA Project
In this Exploratory Data Analysis (EDA) project, we visualized and analyzed various aspects of a dataset related to athletes, their earnings, nationalities, sports, and the corresponding years.
This EDA project provided valuable insights into the dataset, including earnings patterns, sports diversity, nationality-based disparities, and temporal trends. The visualizations and analyses conducted here are crucial for understanding the economic aspects of the sports industry and identifying potential areas for further investigation or modeling.